requirements.txt
FastAPI
uvicorn
transformers
sentence-transformers
faiss-cpu
redis
rq
jinja2
htmx
tailwindcss
PyPDF2
